Improve view selection when fitting images to view
When fitting images to view, we have to select the "smallest" view to
fit the image. All other (larger) views will then automatically have
enough space to show their scene completely.

To do this, we previously chose the smallest view in either the
horizontal or vertical direction, depending on the orientation of the
first image. This did not choose the correct view when the scenes of
different views have different sizes (and different orientations).