Cache transformed versions of images
Since images can now be offset, rotated and colorized, updating the
views can take quite some time to recompute these transformed versions
of the original images. This is not necessary when e.g. only the opacity
ratio of a blending comparison changes. We therefore cache transformed
images until a parameter changes that actually requires them to be
recomputed.